db_perlvdb: Fix incorrect integer downcast
Credits to Aaron Meriwether for spotting the bug and suggesting the fix

Fixes #1952

(cherry picked from commit 3e83412d50b0d8dd5b8a0f7498fb8b66e96ef353)

+++ modules/db_perlvdb/perlvdbfunc.c
@@ -136,7 +136,7 @@ db_con_t* perlvdb_db_init(const str* url) {
 		return NULL;
 	}
 	memset(res, 0, consize);
-	CON_TAIL(res) = (unsigned int)(unsigned long)obj;
+	CON_TAIL(res) = (unsigned long)obj;
 
 	return res;
 }
